+++ config/webpack/loaders/ejs.js
@@ -1,0 +1,4 @@
+module.exports = {
+  test: /\.ejs$/, 
+  loader: 'ejs-compiled-loader'
+}