+++ config/application.rb
@@ -40,11 +40,6 @@ module Fulcrum
       g.test_framework :rspec
     end
 
-    config.webpack.dev_server.host = ENV["WEBPACK_HOST"]
-    config.webpack.dev_server.manifest_host = ENV["WEBPACK_MANIFEST_HOST"]
-    config.webpack.dev_server.manifest_port = ENV["WEBPACK_MANIFEST_PORT"]
-    config.webpack.dev_server.port = ENV["WEBPACK_PORT"]
-
     config.assets.paths << Rails.root.join('node_modules')
   end
 end