+++ config/webpack/loaders/babel.js
@@ -1,0 +1,8 @@
+module.exports = {
+  test: /\.js(\.erb)?$/,
+  exclude: /node_modules/,
+  loader: 'babel-loader',
+  query: {
+    presets: ['airbnb']
+  }
+}