+++ spec/feature_helper.rb
@@ -2,8 +2,8 @@ require 'rails_helper'
 
 RSpec.configure do |config|
   config.before(:suite) do
-    `bundle exec rake assets:precompile`
-    `bundle exec rake webpack:compile`
+    %x[bundle exec rake assets:precompile]
+    Webpacker.compile
   end
 
   config.include Warden::Test::Helpers